Fix sequential MKL case

Summary:
Make sure we don't reset number of OMP threads to 1
in case of sequential MKL library

Test Plan:
USE_CUDA=0 PARALLEL_BACKEND=OPENMP BLAS=MKL USE_MKLDNN=1 MKL_SEQ=1
MKLDNN_THREADING=SEQ BUILD_BINARY=1 python setup.py develop --cmake

./build/bin/parallel_info

gh-metadata: pytorch pytorch 22062 gh/ilia-cher/33/head

+++ aten/src/ATen/native/quantized/cpu/qconv_prepack.cpp
@@ -30,7 +30,7 @@ class QConvPackWeightInt8 final : public c10::OperatorKernel {
     int NDim = output_channels / groups;
     int KDim_per_group = kernel_h * kernel_w * input_channels_per_group;
     auto weight_contig = weight.contiguous();
-    int weight_zero_point_int32 = weight.q_zero_point().toInt();
+    int32_t weight_zero_point_int32 = weight.q_zero_point();
     TORCH_CHECK(
         weight_zero_point_int32 == 0,
         "Only symmetric quantization is supported for weights yet");
@@ -50,7 +50,7 @@ class QConvPackWeightInt8 final : public c10::OperatorKernel {
                              groups),
                          col_offsets,
                          {kernel_h, kernel_w},
-                         weight.q_scale().toFloat(),
+                         weight.q_scale(),
                          weight_zero_point_int32});
     // TODO: we will need to replace this with torchscript classes at a later
     // point.
